@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 2.3.2+27
+version: 2.4.0+28
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.16.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,27 +27,27 @@ dependencies:
     sdk: flutter
   intl: ^0.17.0 
   platform: ^3.1.0
-  font_awesome_flutter: ^9.2.0
-  google_mobile_ads: ^1.1.0
+  font_awesome_flutter: ^10.1.0
+  google_mobile_ads: ^1.2.0
   permission_handler: ^9.2.0
   qr_flutter: ^4.0.0
   clipboard: ^0.1.3
-  share_plus: ^3.1.0
-  url_launcher: ^6.0.20
+  share_plus: ^4.0.4
+  url_launcher: ^6.1.0
   image_gallery_saver: ^1.7.1
   screenshot: ^1.2.3
   shared_preferences: ^2.0.13
   google_fonts: ^2.3.1
-  image_picker: ^0.8.4+11
-  google_ml_kit: ^0.7.3
-  camera: ^0.9.4+16
+  image_picker: ^0.8.5
+  google_mlkit_barcode_scanning: ^0.0.2
+  camera: ^0.9.4+20
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^1.0.4
   flutter_launcher_icons: ^0.9.2
-  flutter_native_splash: ^2.1.0
+  flutter_native_splash: ^2.1.6
 
 flutter_native_splash:
   color: "#ffffff"
