@@ -4,14 +4,9 @@ import 'package:url_launcher/url_launcher.dart';
 mixin OpenLink {
   Future<void> openLink(String url, {VoidCallback? onError}) async {
     try {
-      if (await canLaunch(url)) {
-        await launch(url, forceWebView: false, forceSafariVC: false);
-      } else {
-        if (onError != null) {
-          onError();
-        } else {
-          throw Exception('Error in open link: not implemented error');
-        }
+      if (!await launchUrl(Uri.parse(url),
+          mode: LaunchMode.externalApplication)) {
+        throw Exception('Error in url launcher | not implemented ');
       }
     } catch (error, stackStrace) {
       debugPrint(error.toString());
