@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:google_mlkit_barcode_scanning/google_mlkit_barcode_scanning.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:clipboard/clipboard.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:google_ml_kit/google_ml_kit.dart';
 
 import '../../../../shared/admob/controller/admob_controller.dart';
 import '../../../../shared/admob/widget/native_ad.dart';
@@ -88,7 +88,13 @@ class _ResultReadCodeState extends State<ResultReadCode> with PopupNotices {
                           AppLocalizations.of(context)!.scanResultQrToolTip,
                       onPressed: () =>
                           FlutterClipboard.copy(widget.result).then(
-                        (_) => popupCopyBoard(context),
+                        (_) => popupNotice(
+                          context,
+                          notice: AppLocalizations.of(context)!
+                                  .scanResultPopupCopy +
+                              '.',
+                          duration: const Duration(milliseconds: 500),
+                        ),
                       ),
                       icon: Icon(Icons.copy,
                           color: Theme.of(context).iconTheme.color),
